Move TouchAction classes out of TouchPerformExecutor
They will be used by TouchMultiPerformExecutor as well
